business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/rotator/Dec2Dmm.java
package eu.estcube.gsconnector.commands.rotator;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class Dec2Dmm implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+e ");
        messageString.append(command.getParameter("Dec Deg"));
        messageString.append("\n");
        return messageString;
    }
}
